# public/swagger.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "Badge": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          }
        },
        "required": ["created_at", "updated_at", "name", "slug"],
        "type": "object",
        "additionalProperties": false
      },
      "AppCategoryName": {
        "type": "string",
        "enum": [
          "Uncategorised",
          "Event related",
          "Games",
          "Graphics",
          "Hardware",
          "Utility",
          "Wearable",
          "Data",
          "Silly",
          "Hacking",
          "Troll",
          "Unusable",
          "Adult",
          "Virus",
          "Interpreter"
        ]
      },
      "ProjectStatusName": {
        "type": "string",
        "enum": ["working", "in_progress", "broken", "unknown"]
      },
      "Version": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          },
          "revision": {
            "type": "number",
            "format": "double"
          },
          "semantic_version": {
            "type": "string"
          },
          "zip": {
            "type": "string"
          },
          "size_of_zip": {
            "type": "number",
            "format": "double"
          },
          "git_commit_id": {
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/FileMetadata"
            },
            "type": "array"
          },
          "app_metadata": {
            "$ref": "#/components/schemas/AppMetadataJSON"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "download_count": {
            "type": "number",
            "format": "double"
          },
          "project_slug": {
            "type": "string"
          }
        },
        "required": [
          "created_at",
          "updated_at",
          "revision",
          "files",
          "app_metadata",
          "download_count",
          "project_slug"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "User": {
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "remember_token": {
            "type": "string"
          },
          "editor": {
            "type": "string"
          },
          "public": {
            "type": "boolean"
          },
          "show_projects": {
            "type": "boolean"
          },
          "email_verified_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "created_at",
          "updated_at",
          "id",
          "email",
          "admin",
          "name",
          "password",
          "editor",
          "public",
          "show_projects"
        ],
        "type": "object",
        "additionalProperties": false
      },
      "FileMetadata": {
        "properties": {
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          },
          "editable": {
            "type": "boolean"
          },
          "lintable": {
            "type": "boolean"
          },
          "extension": {
            "type": "string"
          },
          "baseName": {
            "type": "string"
          },
          "mime": {
            "type": "string"
          },
          "size_of_content": {
            "type": "number",
            "format": "double"
          },
          "crc32": {
            "type": "string"
          },
          "size_formatted": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "version",
          "user",
          "created_at",
          "updated_at",
          "extension",
          "baseName",
          "size_of_content",
          "crc32",
          "size_formatted",
          "name"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "ProjectSlug": {
        "type": "string"
      },
      "Dependency": {
        "properties": {
          "project_slug": {
            "$ref": "#/components/schemas/ProjectSlug"
          },
          "semantic_version_range": {
            "type": "string"
          }
        },
        "required": ["project_slug", "semantic_version_range"],
        "type": "object",
        "additionalProperties": false
      },
      "ProjectStatusOnBadge": {
        "properties": {
          "badge": {
            "$ref": "#/components/schemas/Badge"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatusName"
          }
        },
        "required": ["badge", "created_at", "updated_at", "status"],
        "type": "object",
        "additionalProperties": false
      },
      "VoteFromUser": {
        "properties": {
          "type": {
            "type": "string",
            "enum": ["up", "down", "pig"]
          },
          "comment": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["type", "user", "created_at", "updated_at"],
        "type": "object",
        "additionalProperties": false
      },
      "WarningFromUser": {
        "properties": {
          "description": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": ["description", "user", "created_at", "updated_at"],
        "type": "object",
        "additionalProperties": false
      },
      "Project": {
        "properties": {
          "slug": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          },
          "git": {
            "type": "string"
          },
          "allow_team_fixes": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "min_firmware": {
            "type": "number",
            "format": "double"
          },
          "max_firmware": {
            "type": "number",
            "format": "double"
          },
          "git_commit_id": {
            "type": "string"
          },
          "published_at": {
            "type": "string",
            "format": "date-time"
          },
          "download_counter": {
            "type": "number",
            "format": "double"
          },
          "license": {
            "type": "string"
          },
          "size_of_zip": {
            "type": "number",
            "format": "double"
          },
          "size_of_content": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "$ref": "#/components/schemas/AppCategoryName"
          },
          "description": {
            "type": "string"
          },
          "revision": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/ProjectStatusName"
          },
          "author": {
            "type": "string"
          },
          "interpreter": {
            "type": "string"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          },
          "badges": {
            "items": {
              "$ref": "#/components/schemas/Badge"
            },
            "type": "array"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/Dependency"
            },
            "type": "array"
          },
          "states": {
            "items": {
              "$ref": "#/components/schemas/ProjectStatusOnBadge"
            },
            "type": "array"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/Version"
            },
            "type": "array"
          },
          "votes": {
            "items": {
              "$ref": "#/components/schemas/VoteFromUser"
            },
            "type": "array"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningFromUser"
            },
            "type": "array"
          },
          "collaborators": {
            "items": {
              "$ref": "#/components/schemas/User"
            },
            "type": "array"
          }
        },
        "required": ["slug", "user_id", "created_at", "updated_at", "category"],
        "type": "object",
        "additionalProperties": false
      },
      "Uint8Array": {
        "description": "A typed array of 8-bit unsigned integer values. The contents are initialized to 0. If the\nrequested number of bytes could not be allocated an exception is raised.",
        "properties": {},
        "type": "object",
        "additionalProperties": false
      },
      "Pick_DBInsertUser.Exclude_keyofDBInsertUser.id__": {
        "properties": {
          "email": {
            "type": "string"
          },
          "admin": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "remember_token": {
            "type": "string"
          },
          "editor": {
            "type": "string"
          },
          "public": {
            "type": "boolean"
          },
          "show_projects": {
            "type": "boolean"
          },
          "email_verified_at": {
            "type": "string"
          }
        },
        "required": ["email", "name", "password"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_DBInsertUser.id_": {
        "$ref": "#/components/schemas/Pick_DBInsertUser.Exclude_keyofDBInsertUser.id__",
        "description": "Construct a type with the properties of T except for those in type K."
      },
      "Pick_DBInsertProject.Exclude_keyofDBInsertProject.slug__": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/AppCategoryName"
          },
          "version_id": {
            "type": "number",
            "format": "double"
          },
          "git": {
            "type": "string"
          },
          "allow_team_fixes": {
            "type": "boolean"
          },
          "user_id": {
            "type": "string"
          }
        },
        "required": ["user_id"],
        "type": "object",
        "description": "From T, pick a set of properties whose keys are in the union K"
      },
      "Omit_DBInsertProject.slug_": {
        "$ref": "#/components/schemas/Pick_DBInsertProject.Exclude_keyofDBInsertProject.slug__",
        "description": "Construct a type with the properties of T except for those in type K."
      },
      "Partial_Omit_DBProject.slug__": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/AppCategoryName"
          },
          "version_id": {
            "type": "number",
            "format": "double"
          },
          "git": {
            "type": "string"
          },
          "allow_team_fixes": {
            "type": "boolean"
          },
          "user_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "deleted_at": {
            "type": "string"
          }
        },
        "type": "object",
        "description": "Make all properties in T optional"
      }
    },
    "securitySchemes": {}
  },
  "info": {
    "title": "badgehub-api",
    "version": "3",
    "description": "Node project for the BadgeHub API",
    "license": {
      "name": "MIT"
    },
    "contact": {}
  },
  "paths": {
...
    "/api/v3/apps": {
      "get": {
        "operationId": "GetProjects",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  },
                  "type": "array"
                }
              }
            }
          }
        },
        "description": "Get list of apps, optionally limited by page start/length and/or filtered by category",
        "tags": ["public"],
        "security": [],
        "parameters": [
          {
            "in": "query",
            "name": "pageStart",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "pageLength",
            "required": false,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "device",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v3/apps/{slug}": {
      "get": {
        "operationId": "GetProject",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "reason": {
                      "type": "string"
                    }
                  },
                  "required": ["reason"],
                  "type": "object"
                }
              }
            }
          }
        },
        "description": "Get app details",
        "tags": ["public"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "InsertProject",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "description": "Create a new app",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectSlug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Omit_DBInsertProject.slug_"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "DeleteProject",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "description": "Create a new app",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectSlug"
            }
          }
        ]
      },
      "patch": {
        "operationId": "UpdateProject",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "description": "Create a new app",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ProjectSlug"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Partial_Omit_DBProject.slug__"
              }
            }
          }
        }
      }
    },
    "/api/v3/apps/{slug}/files/latest/{filePath}": {
      "get": {
        "operationId": "GetLatestPublishedFile",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Uint8Array"
                }
              }
            }
          }
        },
        "description": "get the latest published version of a file in the project",
        "tags": ["public"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filePath",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v3/apps/{slug}/files/rev{revision}/{filePath}": {
      "get": {
        "operationId": "GetFileForVersion",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Uint8Array"
                }
              }
            }
          }
        },
        "description": "get a file for a specific version of the project",
        "tags": ["public"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          },
          {
            "in": "path",
            "name": "filePath",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v3/apps/{slug}/zip/latest": {
      "get": {
        "operationId": "GetLatestPublishedZip",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Uint8Array"
                }
              }
            }
          }
        },
        "description": "get the latest published version of the app in zip format",
        "tags": ["public"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v3/apps/{slug}/zip/rev{revision}": {
      "get": {
        "operationId": "GetZipForVersion",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Uint8Array"
                }
              }
            }
          }
        },
        "description": "get the app zip for a specific version of the project",
        "tags": ["public"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "revision",
            "required": true,
            "schema": {
              "format": "double",
              "type": "number"
            }
          }
        ]
      }
    },
    "/api/v3/users/{userId}": {
      "post": {
        "operationId": "InsertUser",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "description": "Create a new user",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Omit_DBInsertUser.id_"
              }
            }
          }
        }
      }
    },
    "/api/v3/apps/{slug}/files/draft/{filePath}": {
      "post": {
        "operationId": "WriteFile",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "description": "Upload a file to the latest draft version of the project.",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filePath",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "string"
                  },
                  {
                    "$ref": "#/components/schemas/Uint8Array"
                  }
                ]
              }
            }
          }
        }
      },
      "get": {
        "operationId": "GetDraftFile",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Uint8Array"
                }
              }
            }
          }
        },
        "description": "get the latest draft version of the project.",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "filePath",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v3/apps/{slug}/zip/draft": {
      "get": {
        "operationId": "GetLatestPublishedZip",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Uint8Array"
                }
              }
            }
          }
        },
        "description": "get the latest draft version of the app in zip format",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v3/apps/{slug}/zip": {
      "post": {
        "operationId": "WriteZip",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Version"
                }
              }
            }
          }
        },
        "description": "Upload a file to the latest draft version of the project.",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Uint8Array"
              }
            }
          }
        }
      }
    },
    "/api/v3/apps/{slug}/publish": {
      "patch": {
        "operationId": "PublishVersion",
        "responses": {
          "204": {
            "description": "No content"
          }
        },
        "description": "Publish the latest draft version",
        "tags": ["private"],
        "security": [],
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    }
  },
  "servers": [
    {
      "url": "https://localhost:8081/"
    }
  ]
}